added separate new GameButton,LobbyPane,MenuItemPainter and its implementation, changed StartMenuView, completed CardModelViewTest
