+++ app/src/main/java/org/abianchi/dubito/app/App.java
@@ -2,8 +2,16 @@ package org.abianchi.dubito.app;
 
 import org.abianchi.dubito.app.views.*;
 
+import java.awt.*;
+
 public class App {
     public static void main(final String[] args) {
-        new StartMenuView();
+        EventQueue.invokeLater(new Runnable() {
+            @Override
+            public void run() {
+                new StartMenuView();
+            }
+        });
+
     }
 }
