+++ app/src/main/java/org/abianchi/dubito/app/views/LobbyPane.java
@@ -1,0 +1,7 @@
+package org.abianchi.dubito.app.views;
+
+import javax.swing.*;
+
+class LobbyPane extends JPanel {
+
+}
